+++ src/collapse/collapse-panel.vue
@@ -1,31 +1,18 @@
 <template>
-  <div ref="wrapDOM" :class="className">
-    <div ref="headDOM" :class="`${baseClass}__header`" @click="(e) => onChange(e, 'header')">
-      <div :class="`${baseClass}__title`">
-        <slot name="title">{{ title }}</slot>
+  <div ref="$wrap" :class="className">
+    <div ref="$head" :class="`${classPrefix}__header`" @click="handleClick">
+      <div :class="`${classPrefix}__title`">
+        <slot name="header">{{ header }}</slot>
       </div>
-      <div :class="`${baseClass}__header-right`" @click="onChange">
-        <div v-if="extra || $slots.extra" :class="`${baseClass}__header-extra`">
-          <slot name="extra">{{ extra }}</slot>
+      <div :class="`${classPrefix}__header-right`">
+        <div v-if="headerRightContent || $slots.extra" :class="`${classPrefix}__header-extra`">
+          <slot name="headerRightContent">{{ headerRightContent }}</slot>
         </div>
-        <component :is="rightIcon" :class="`${baseClass}__header-icon`"> </component>
+        <component :is="rightIcon" :class="`${classPrefix}__header-icon`" />
       </div>
     </div>
-    <div ref="bodyDOM" :class="`${baseClass}__body`">
-      <div v-if="$slots.default" :class="`${baseClass}__body-slot`">
-        <slot></slot>
-      </div>
-      <template v-else>
-        <div v-for="(c, i) in contList" :key="i" :class="contentClassName(c)">
-          <template v-if="typeof c === 'object'">
-            <div :class="`${baseClass}-list__label`" :style="listLabelStyle">
-              {{ c.label }}
-            </div>
-            <div :class="`${baseClass}-list__content`">{{ c.content }}</div>
-          </template>
-          <template v-else>{{ c }}</template>
-        </div>
-      </template>
+    <div ref="$body" :class="`${classPrefix}__body`">
+      <t-node :content="panelContent"></t-node>
     </div>
   </div>
 </template>
@@ -33,8 +20,6 @@
 <script lang="ts">
 import {
   ref,
-  reactive,
-  toRefs,
   computed,
   nextTick,
   watch,
@@ -42,82 +27,73 @@ import {
   inject,
   SetupContext,
   defineComponent,
+  getCurrentInstance,
 } from 'vue';
 import { ChevronDownIcon, ChevronUpIcon } from 'tdesign-icons-vue-next';
-import {
-  CollapsePropsType,
-  CollapseStateType,
-  CollapsePanelPropsType,
-  CollapsePanelProps,
-  onChangeEvent,
-} from './collapse.interface';
+import props from './collapse-panel-props';
 import config from '../config';
-import { findIndex, isFalsy, toArray } from './util';
-import { useEmitEvent } from '../shared';
+import { findIndex } from './util';
+import { renderTNode, renderContent, TNode } from '../shared';
+import { CollapseProvide } from './collapse.vue';
 
 const { prefix } = config;
 const name = `${prefix}-collapse-panel`;
 export default defineComponent({
   name,
-  components: { ChevronDownIcon, ChevronUpIcon },
-  props: CollapsePanelProps,
-  emits: ['click'],
-  setup(props: CollapsePanelPropsType, context: SetupContext) {
-    const emitEvent = useEmitEvent(props, context.emit);
-    // 从父组件取属性、状态和控制函数
-    const collapseProps = inject('collapseProps') as CollapsePropsType;
-    const collapseState = inject('collapseState') as CollapseStateType;
-    const onPanelChange = inject('onPanelChange') as onChangeEvent;
-    const rightIcon = computed(() => (isActive.value ? ChevronDownIcon : ChevronUpIcon));
-    // 内容转为数组统一处理
-    const contList = computed(() => toArray(props.content));
+  components: { ChevronDownIcon, ChevronUpIcon, TNode },
+  props,
+  setup(props, context: SetupContext) {
+    const internalInstance = getCurrentInstance();
+    const parent = inject<CollapseProvide>('collapse');
+    const isTrue = (val: any) => typeof val === 'boolean' && val;
+
+    const rightIcon = computed(() => {
+      if (props.expandIcon === false) return;
+      if (isTrue(props.expandIcon) || isTrue(parent?.expandIcon.value)) {
+        console.log(isTrue(parent?.expandIcon.value));
+
+        return isActive.value ? ChevronDownIcon : ChevronUpIcon;
+      }
+      return renderTNode(internalInstance, 'expand-icon')[0];
+    });
+    const disabled = computed(() => parent?.disabled.value || props.disabled);
     const className = computed(() => ({
       [`${name}`]: true,
       [`${name}--active`]: isActive.value,
-      [`${name}--disabled`]: props.disabled,
+      [`${name}--disabled`]: disabled.value,
     }));
-    const labelWidth = computed(() => props.labelWidth || collapseProps.labelWidth);
-    const contentClassName = computed(() => (c: number | string | Record<string, unknown>) => [
-      `${name}__content`,
-      typeof c === 'object' ? `${name}-list__item` : '',
-    ]);
-    const listLabelStyle = computed(() => (!isFalsy(labelWidth.value) ? { width: `${labelWidth.value}px` } : {}));
-    // 是否展开态
-    const isActive = computed(() => findIndex(props.name, collapseState.curValue) > -1);
-    const state = reactive({
-      baseClass: name,
-    });
-
-    // 切换自身展开态
-    const onChange = (e: any = null, from = '') => {
-      e?.stopPropagation();
-      if (props.disabled) {
-        return;
+    const isActive = computed(() => findIndex(props.value, parent?.activeValue.value) > -1);
+    const updatePanelValue = () => {
+      if (props.value != null) {
+        parent?.onPanelChange(props.value);
       }
-      emitEvent('click', props.name);
-      if (/^header$/i.test(from) && !props.headerClickable) {
-        return;
-      }
-      onPanelChange(props.name);
     };
+    const handleClick = (e: MouseEvent) => {
+      e?.stopPropagation();
+      if (disabled.value) {
+        return;
+      }
+      updatePanelValue();
+    };
+    const panelContent = renderContent(internalInstance, 'default', 'content');
 
     // 设置折叠/展开高度过渡
-    const bodyDOM = ref();
-    const wrapDOM = ref();
-    const headDOM = ref();
+    const $body = ref();
+    const $wrap = ref();
+    const $head = ref();
     const updatePanelState = () => {
-      if (!wrapDOM.value) {
-        console.log('[collapse] 组件尚未挂载', wrapDOM.value);
+      if (!$wrap.value) {
+        // console.log('[collapse] 组件尚未挂载', $wrap.value);
         return;
       }
-      const { height: headHeight } = headDOM.value.getBoundingClientRect();
+      const { height: headHeight } = $head.value.getBoundingClientRect();
       if (!isActive.value) {
-        wrapDOM.value.style.height = `${headHeight}px`;
+        $wrap.value.style.height = `${headHeight}px`;
         return;
       }
-      const { height: bodyHeight } = bodyDOM.value.getBoundingClientRect();
+      const { height: bodyHeight } = $body.value.getBoundingClientRect();
       const height = headHeight + bodyHeight;
-      wrapDOM.value.style.height = `${height}px`;
+      $wrap.value.style.height = `${height}px`;
     };
 
     watch(isActive, () => {
@@ -125,24 +101,23 @@ export default defineComponent({
         updatePanelState();
       });
     });
+
     onMounted(() => {
-      if (collapseProps.expandAll) {
-        // 展开未展开的
-        !isActive.value && onChange();
+      if (parent?.defaultExpandAll) {
+        updatePanelValue();
       }
       updatePanelState();
     });
+
     return {
-      ...toRefs(state),
-      contList,
-      headDOM,
+      classPrefix: name,
       rightIcon,
-      bodyDOM,
-      wrapDOM,
+      $head,
+      $body,
+      $wrap,
       className,
-      contentClassName,
-      listLabelStyle,
-      onChange,
+      handleClick,
+      panelContent,
     };
   },
 });