+++ src/notice-bar/demos/icon.vue
@@ -1,38 +1,22 @@
 <template>
   <div class="noticebar-demo">
-    <t-cell-group>
-      <t-cell value-align="left">
-        <t-notice-bar content="提示文字描述提示文字描述提示文字描述" />
-        <t-notice-bar
-          content="提示文字描述提示文字描述提示文字描述提示文字描述提示文字描述提示文字描述提示文字描述提示文字描述提示文字描述提示文字描述"
-        />
-        <t-notice-bar type="error" content="提示文字描述提示文字描述提示文字描述" />
-        <t-notice-bar
-          type="error"
-          content="提示文字描述提示文字描述提示文字描述提示文字描述提示文字描述提示文字描述提示文字描述提示文字描述提示文字描述提示文字描述"
-        />
-      </t-cell>
-    </t-cell-group>
+    <tdesign-demo-block summary="带图标静态公告栏">
+      <t-notice-bar visible content="提示文字描述提示文字描述提示文字描述" />
+    </tdesign-demo-block>
   </div>
 </template>
 
 <script lang="ts">
-import { defineComponent } from 'vue';
+import { reactive, defineComponent, toRefs, h } from 'vue';
 
 export default defineComponent({
+  components: {},
   setup() {
-    return {};
+    const state = reactive({});
+
+    return {
+      ...toRefs(state),
+    };
   },
 });
 </script>
-
-<style lang="less" scoped>
-.noticebar-demo {
-  background-color: #f5f5f5;
-
-  .t-cell {
-    background-color: transparent;
-    padding: 0;
-  }
-}
-</style>