+++ src/notice-bar/notice-bar.vue
@@ -1,32 +1,30 @@
 <template>
-  <div :class="rootClasses" :style="bgColorCustom">
+  <div v-if="isShow" :class="rootClasses">
     <div :class="`${name}__inner`">
-      <div v-if="computedLeftIcon !== undefined" :class="`${name}__hd`">
-        <t-node :content="computedLeftIcon" :style="iconColorCustom"></t-node>
+      <div v-if="computedPrefixIcon !== undefined" :class="`${name}__hd`" @click="() => handleClick('prefix-icon')">
+        <t-node :content="computedPrefixIcon"></t-node>
       </div>
+
       <div :class="`${name}__bd`">
-        <div ref="listDOM" :class="`${name}__list ${scrollable ? `${name}__list--scrolling` : ''}`">
+        <div ref="listDOM" :class="`${name}__list ${scroll.marquee ? `${name}__list--scrolling` : ''}`">
           <div
             ref="itemDOM"
-            :class="`${name}__item ${showDetailText ? `${name}__item-detail` : ''}`"
-            :style="scrollable ? animateStyle : ''"
+            :class="`${name}__item ${showExtraText ? `${name}__item-detail` : ''}`"
+            :style="scroll.marquee ? animateStyle : ''"
             @transitionend="handleTransitionend()"
           >
-            <span :class="`${name}__text`" :style="colorCustom"
-              >{{ content
-              }}<span
-                v-if="showDetailText"
-                :class="`${name}__text-detail`"
-                :style="colorCustom"
-                @click="handleDetailLink"
-                >{{ detailText }}</span
-              >
+            <span :class="`${name}__text`" @click="() => handleClick('content')">
+              {{ content }}
+              <span v-if="showExtraText" :class="`${name}__text-detail`" @click.stop="() => handleClick('extra')">
+                {{ extra }}
+              </span>
             </span>
           </div>
         </div>
       </div>
-      <div v-if="computedRightIcon !== undefined" :class="`${name}__ft`" @click="handleClickIcon">
-        <t-node :content="computedRightIcon" :style="iconColorCustom"></t-node>
+
+      <div v-if="computedSuffixIcon !== undefined" :class="`${name}__ft`" @click="() => handleClick('suffix-icon')">
+        <t-node :content="computedSuffixIcon"></t-node>
       </div>
     </div>
   </div>
@@ -44,24 +42,32 @@ import {
   defineComponent,
   h,
   getCurrentInstance,
+  watch,
 } from 'vue';
-import { CloseIcon, ChevronRightIcon } from 'tdesign-icons-vue-next';
+import { InfoCircleFilledIcon, CheckCircleFilledIcon, CloseCircleFilledIcon } from 'tdesign-icons-vue-next';
 import NoticeBarProps from './props';
+import { NoticeBarTrigger, DrawMarquee } from './type';
 
 import config from '../config';
-import { useEmitEvent, renderTNode, TNode } from '../shared';
+import { useEmitEvent, renderTNode, TNode, useVModel } from '../shared';
 
 const { prefix } = config;
 const name = `${prefix}-notice-bar`;
-
+const iconDefault = {
+  info: h(InfoCircleFilledIcon),
+  success: h(CheckCircleFilledIcon),
+  warning: h(InfoCircleFilledIcon),
+  error: h(CloseCircleFilledIcon),
+};
 export default defineComponent({
   name,
   components: { TNode },
   props: NoticeBarProps,
-  emits: ['click', 'close', 'detail'],
+  emits: ['click', 'change'],
   setup(props, context: SetupContext) {
     const emitEvent = useEmitEvent(props, context.emit);
     const internalInstance = getCurrentInstance();
+    // 初始化数据
     const state = reactive({
       duration: 0,
       offset: 0,
@@ -69,47 +75,34 @@ export default defineComponent({
       itemWidth: 0,
       timer: null,
       nextTimer: null,
+      scroll: {
+        marquee: false,
+        speed: 50,
+        loop: -1, // 值为 -1 表示循环播放，值为 0 表示不循环播放
+        delay: 0,
+      },
     });
 
-    const rootClasses = computed(() => [`${name}`, `${name}--info`]);
-    const iconType = {
-      link: ChevronRightIcon,
-      closeable: CloseIcon,
-    };
-    const computedLeftIcon = computed(() => renderTNode(internalInstance, 'leftIcon'));
-    const computedRightIcon = computed(() => {
-      let rightIcon = renderTNode(internalInstance, 'rightIcon');
-      if (rightIcon === undefined) {
-        if (props.mode !== undefined) {
-          rightIcon = h(iconType[props.mode]);
-        }
+    const rootClasses = computed(() => [`${name}`, `${name}--${props.theme}`]);
+    let computedPrefixIcon: any;
+    if (props.prefixIcon !== '') {
+      if (Object.keys(iconDefault).includes(props?.theme as string)) {
+        const key = props.theme as string;
+        computedPrefixIcon = computed(() => iconDefault?.[key]);
       }
-      return rightIcon;
-    });
-    const showDetailText = computed(() => props.detailText !== '');
-    const colorCustom = computed(() => (props.color ? `color:${props.color}` : ''));
-    const bgColorCustom = computed(() => (props.bgColor ? `background-color:${props.bgColor}` : ''));
-    const iconColorCustom = computed(() => (props.iconColor ? `color:${props.iconColor};` : ''));
-
-    function handleClose() {
-      emitEvent('close');
+      computedPrefixIcon = props.prefixIcon
+        ? computed(() => renderTNode(internalInstance, 'prefixIcon'))
+        : computedPrefixIcon;
     }
-
-    function handleClick() {
-      emitEvent('click');
+    // suffix-icon
+    const computedSuffixIcon = computed(() => renderTNode(internalInstance, 'suffixIcon'));
+    // extra
+    const showExtraText = computed(() => renderTNode(internalInstance, 'extra'));
+    // click
+    function handleClick(trigger: NoticeBarTrigger) {
+      emitEvent('click', trigger);
     }
-
-    const handleClickIcon = computed(() => {
-      if (props.mode === 'closeable') {
-        return handleClose;
-      }
-      return handleClick;
-    });
-
-    function handleDetailLink() {
-      emitEvent('detail');
-    }
-
+    // 动画
     const animateStyle = computed(() => ({
       transform: state.offset ? `translateX(${state.offset}px)` : '',
       transitionDuration: `${state.duration}s`,
@@ -119,59 +112,90 @@ export default defineComponent({
     const listDOM = ref();
     const itemDOM = ref();
 
+    const { visible, modelValue } = toRefs(props);
+    const [isShow, setStatusValue] = useVModel(
+      visible,
+      modelValue,
+      props.defaultVisible,
+      props.onChange as (value: boolean | undefined) => void,
+    );
     function handleScrolling() {
-      const delay = props.delay && props.delay > 0 ? props.delay * 1000 : 0;
-      const speed = props.speed && props.speed > 0 ? props.speed : 50;
-
+      if (!props?.marquee || (props?.marquee as DrawMarquee)?.loop === 0) {
+        return;
+      }
+      // 初始化动画参数
+      if (typeof props.marquee === 'boolean') {
+        state.scroll = { ...state.scroll, marquee: props.marquee };
+      }
+      const marquee = props.marquee as DrawMarquee;
+      state.scroll = {
+        marquee: true,
+        loop: typeof marquee?.loop === 'undefined' ? state.scroll.loop : marquee.loop,
+        speed: marquee.speed ?? state.scroll.speed,
+        delay: marquee.delay ?? state.scroll.delay,
+      };
+      // 设置动画
       setTimeout(() => {
-        if (!props.scrollable) {
-          return;
-        }
-
-        // FIXME: getBoundingClientRect报错问题
         const listDOMWidth = listDOM.value?.getBoundingClientRect().width;
         const itemDOMWidth = itemDOM.value?.getBoundingClientRect().width;
-
         if (itemDOMWidth > listDOMWidth) {
           state.offset = -itemDOMWidth;
-          state.duration = itemDOMWidth / speed;
+          state.duration = itemDOMWidth / state.scroll.speed;
           state.listWidth = listDOMWidth;
           state.itemWidth = itemDOMWidth;
         }
-      }, delay);
+      }, state.scroll.delay);
     }
-
+    // 动画结束后，初始化动画
     function handleTransitionend() {
-      const speed = props.speed && props.speed > 0 ? props.speed : 50;
-
+      // 触发再次滚的
+      state.scroll.loop = --state.scroll.loop;
+      if (state.scroll.loop === 0) {
+        state.scroll = {
+          ...state.scroll,
+          marquee: false,
+        };
+        return;
+      }
       state.offset = state.listWidth;
       state.duration = 0;
 
       setTimeout(() => {
         state.offset = -state.itemWidth;
-        state.duration = (state.itemWidth + state.listWidth) / speed;
+        state.duration = (state.itemWidth + state.listWidth) / state.scroll.speed;
       }, 0);
     }
-
     onMounted(() => {
       nextTick(() => {
-        handleScrolling();
+        if (isShow.value) {
+          handleScrolling();
+        }
       });
     });
+    watch(
+      () => isShow.value,
+      () => {
+        emitEvent('change', isShow.value);
+        nextTick(() => {
+          if (isShow.value) {
+            state.offset = state.listWidth;
+            state.duration = 0;
+            handleScrolling();
+          }
+        });
+      },
+    );
 
     return {
       name,
       ...toRefs(props),
       ...toRefs(state),
       rootClasses,
-      computedLeftIcon,
-      computedRightIcon,
-      showDetailText,
-      colorCustom,
-      bgColorCustom,
-      iconColorCustom,
-      handleClickIcon,
-      handleDetailLink,
+      computedPrefixIcon,
+      computedSuffixIcon,
+      showExtraText,
+      isShow,
+      handleClick,
       listDOM,
       itemDOM,
       animateStyle,