+++ src/notice-bar/type.ts
@@ -8,71 +8,60 @@ import { TNode } from '../common';
 
 export interface TdNoticeBarProps {
   /**
-   * 背景色
-   * @default ''
-   */
-  bgColor?: string;
-  /**
-   * 文本颜色
-   * @default ''
-   */
-  color?: string;
-  /**
    * 文本内容
-   * @default ''
    */
-  content?: string;
+  content?: string | TNode;
   /**
-   * 延迟显示，单位：毫秒
-   * @default 0
+   * 右侧额外信息
    */
-  delay?: number;
+  extra?: string | TNode;
   /**
-   * 详情
-   * @default ''
+   * 跑马灯效果。speed 指速度控制；loop 指循环播放次数，值为 -1 表示循环播放，值为 0 表示不循环播放；delay 表示延迟多久开始播放
+   * @default false
    */
-  detailText?: string;
+  marquee?: boolean | DrawMarquee;
   /**
-   * 详情颜色
-   * @default ''
+   * 前缀图标
    */
-  detailTextColor?: string;
+  prefixIcon?: string | TNode;
   /**
-   * 图标颜色
-   * @default ''
+   * 后缀图标
    */
-  iconColor?: string;
+  suffixIcon?: TNode;
   /**
-   * 左边图标
+   * 内置主题
+   * @default info
    */
-  leftIcon?: TNode;
+  theme?: 'info' | 'success' | 'warning' | 'error';
   /**
-   * 模式
+   * 显示/隐藏
+   * @default false
    */
-  mode?: 'link' | 'closeable';
+  visible?: boolean;
   /**
-   * 左边图标
+   * 显示/隐藏，非受控属性
+   * @default false
    */
-  rightIcon?: TNode;
+  defaultVisible?: boolean;
   /**
-   * 是否需要滚动
+   * 显示/隐藏
+   * @default false
    */
-  scrollable?: boolean;
+  modelValue?: boolean;
   /**
-   * 滚动速度
-   * @default 50
+   * 展示或关闭公告栏时触发。参数为true时，代表展示公告栏。参数为false时，代表关闭公告栏
    */
-  speed?: number;
+  onChange?: (value: boolean) => void;
   /**
    * 点击事件
    */
-  onClick?: () => void;
-  /**
-   * 关闭
-   */
-  onClose?: () => void;
-  /**
-   * 详情点击
-   */
-  onDetail?: () => void;
+  onClick?: (trigger: NoticeBarTrigger) => void;
 }
+
+export interface DrawMarquee {
+  speed?: number;
+  loop?: number;
+  delay?: number;
+}
+
+export type NoticeBarTrigger = 'prefix-icon' | 'content' | 'extra' | 'suffix-icon';