+++ src/swipe-cell/type.ts
@@ -41,6 +41,15 @@ export interface TdSwipeCellProps {
   onClick?: (context: SwipeActionClickContext) => void;
 }
 
-export interface SwipeActionItem { text: string; className?: string; style?: string; onClick?: () => void; [key: string]: any };
+export interface SwipeActionItem {
+  text: string;
+  className?: string;
+  style?: string;
+  onClick?: () => void;
+  [key: string]: any;
+}
 
-export interface SwipeActionClickContext { action: SwipeActionItem; source: 'left' | 'right' };
+export interface SwipeActionClickContext {
+  action: SwipeActionItem;
+  source: 'left' | 'right';
+}